Big update!
Bunch of changes in the create\edit part (Not started on "clone" but in theory all that's missing is 1 line of code); Created directive for all fields in the form; Which makes it "automagicly" work only by adding a reference to the field in "schemes".

Fixed the layout, to reflect the changes we discussed (Removing minified list); Menu is now on right-hand side (After hours of testing, I decided I don't have time\patience\knowledge to pull off a custom top-bar menu)

Implicit in the changes to the list\layout: List is now it's own jade file; Changes the search-box for the list.jade page to follow the same look as the menu; Did the same to the add\edit page; Also hooked up resolve to the listroute; That means we _know_ data is loaded before the page is loaded. That eliminated the uncertanty with asking for the data after the user accesses the page.

moar moar updates:

Added a "contextualMenu"; It's quite simple so I'm hoping we can extend is somewhat; What it does is present different options based on the location of the user.

Added a bunch of more data to be resolved to a add\edit\clone route, as it has a bunch of hidden dependencies in the api (3-4 different api calls just to get the groups\datasets\etc to display the basic form)

TLDR;
Fuck im tired!