dataElementService.getElement() now utilizes the getAllElements() method to get the requested item;
The reason is that we think it's better to get all of the waiting\loading done before the app is further used; This way there will only be minimal loading later in the app. This is only a downside if the app is initialized on a route requesting a  single element; In normal, getAllElements() would be used anyways.

 Also fixed route.resolve for clone to remove ID of the cloned object, so it's truly a NEW clone, and not the exact same element.